Exception handling - delete

+++ src/main/java/com/nelio/course/services/UserService.java
@@ -4,11 +4,14 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.nelio.course.entities.User;
 
 import com.nelio.course.repositories.UserRepository;
+import com.nelio.course.services.exceptions.DatabaseException;
 import com.nelio.course.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -32,8 +35,17 @@ public class UserService {
     }
 
     public void delete(Long id){
-        repository.deleteById(id);
+        if (findById(id) == null) {
+            throw new EmptyResultDataAccessException("object don't exist in the database", findAll().size());                 
+        }  
+        try {
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }  
+        
     }
+
 
     public User update(long id, User obj){
         User entity = repository.getReferenceById(id);
